fix: 로그를 위해 생성한 익명 클래스 BeanPostProcessor를 별도의 클래스로 분리
